+++ branch-alpha/sqlmath_wasm.js
@@ -1,7 +1,7 @@
 /*jslint-disable*/
 // Copyright (c) 2021 Kai Zhu
 // SPDX-License-Identifier: MIT
-// 2025-09-28T06:55:57+0000
+// 2025-09-28T09:51:17+0000
 (function () {
 "use strict";